Dynamic Crosshair Implementation
+A new dynamic crosshair accuracy implementation has been added! While the crosshair isn't perfect by any stretch (No limits to how unnacurate a gun can get and it (Crosshair) isn't 100% accurate itself it's a very close approximation)

#Weapon Balance changes.
Shotgun fires quicker in between blasts to make it feel more responsive

Shots are tighter and closer to the center however firing over and over will be innacurate.

The shotguns per bullet damage has been decreased intensely to make up for the increase in fire rate.

#Rifle
The rifle has not seen many changes other than clip size and another damage reduction. I believe this is more fun for the player as they can fire a weapon for longer and really see the affects of overfiring with the new accuracy
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~